invader timestamp

+++ include/iqzz.h
@@ -1,4 +1,4 @@
 #pragma once
 
 #include <stdint.h>
-void quantification_inverse(double *dct_coeff, const uint8_t *quantification_table);
+void quantification_inverse(int *dct_coeff,const uint16_t *quantification_table);
